requirements.txt
aiogram==3.13.1
python_dotenv == 1.0.1
